third_party/WebKit/LayoutTests/external/wpt/referrer-policy/strict-origin-when-cross-origin/attr-referrer/same-origin/http-https/script-tag/upgrade-protocol.swap-origin-redirect.http-expected.txt

This is a testharness.js-based test.
FAIL The referrer URL is origin when a
                                 document served over http requires an https
                                 sub-resource via script-tag using the attr-referrer
                                 delivery method with swap-origin-redirect and when
                                 the target request is same-origin. assert_equals: Reported Referrer URL is 'origin'. expected "http://web-platform.test:8001/" but got "http://web-platform.test:8001/referrer-policy/strict-origin-when-cross-origin/attr-referrer/same-origin/http-https/script-tag/upgrade-protocol.swap-origin-redirect.http.html"
Harness: the test ran to completion.
